wip: implement concurrent async indexing of different chats

+++ src/dotnet/MLSearch.Service/Engine/OpenSearch/Setup/ClusterSetup.cs
@@ -273,7 +273,7 @@ internal class ClusterSetup(
                                         "type": "object",
                                         "properties": {
                                             "{{chatSliceEntryIdField}}":  { "type": "keyword" },
-                                            "{{chatSliceEntryVersionField}}": { "type": "integer" }
+                                            "{{chatSliceEntryVersionField}}": { "type": "long" }
                                         }
                                     },
                                     "{{startOffsetField}}": { "type": "integer" },
